theme: Overide date read-time and no excerpt in overlay
Modify the archive-single.html and single.html to display the date and
the read-time in a single line of "meta" data. Always display this in
the single layout, regardless if overlay is used or not.

Never include the excerpt in the overlay, just the title. Excerpt has at
least two uses and should probably be split into overlay_excerpt and
archive_excerpt so they can be configured independently.

Signed-off-by: Darren Hart <darren@dvhart.com>